pyproject.toml
[project]
name = "protocol-aura"
version = "0.1.0"
description = "Agent-to-Agent Retail Layer for Agentic Commerce - Negotiation over meaning, not metadata"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "streamlit>=1.40.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "google-generativeai>=0.8.0",
    "numpy>=2.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.5.0",
    "plotly>=5.24.0",
]

[project.scripts]
aura-api = "protocol_aura.api.main:run"
aura-ui = "protocol_aura.ui.dashboard:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/protocol_aura"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
